feat: separate loader file
